feat(graphql): accept amount filters above the 32-bit limit (#4155)
The backend moved `amountFrom` / `amountTo` from `Int` to `BigInt`
(invoices, credit notes, and the two data-export filter inputs). Our two
query documents still declared those variables as `Int`, which no longer
validates against the new schema.

- `getInvoicesList` and `getCreditNotesList` now declare `$amountFrom` /
`$amountTo` as `BigInt`.
- `src/generated/graphql.tsx` regenerated with `pnpm codegen`.

No logic change: `BigInt` already maps to `Scalars['BigInt']` (`any`),
so no call site moves.

**Merge this after the backend schema change lands.** Until then the
deployed API still declares these arguments as `Int` and would reject a
query declaring them as `BigInt`.

### The regenerated file also picks up unrelated pending drift

Codegen is whole-file, and the committed types were behind the schema on
three points that have nothing to do with amount filters:

- `Plan.interval`, `Plan.payInAdvance` and `Plan.amountCents` become
nullable. #4152 already adapted the call sites for a nullable interval
but did not regenerate the types, so this is the catch-up.
- `FeatureFlagEnum` loses `MultiCurrency` and `MultiEntityBilling`.
Nothing in the app references either.
- The `BigInt` scalar doc comment gets re-wrapped.

Trimming those back out would mean hand-editing a generated file, so
they are left exactly as codegen produced them. That accounts for most
of the diff in `graphql.tsx`; the amount-filter change itself is 12
lines there plus the 4 in the query documents.

No test is added: this is a type-level regeneration with no runtime
behavior to assert, and the typecheck is the gate. Lint, types,
translations and the full Jest suite (920 suites / 11344 tests) are
green.

+++ src/pages/CreditNotesPage.tsx
@@ -34,8 +34,8 @@ import { useOrganizationInfos } from '~/hooks/useOrganizationInfos'
 
 gql`
   query getCreditNotesList(
-    $amountFrom: Int
-    $amountTo: Int
+    $amountFrom: BigInt
+    $amountTo: BigInt
     $creditStatus: [CreditNoteCreditStatusEnum!]
     $currency: CurrencyEnum
     $customerExternalId: String
